Fix issues caused by wrong refactoring
