+++ subprojects/testing-jvm-infrastructure/src/main/java/org/gradle/api/internal/tasks/testing/failure/mappers/AssertErrorMapper.java
@@ -16,43 +16,30 @@
 
 package org.gradle.api.internal.tasks.testing.failure.mappers;
 
+import org.gradle.api.NonNullApi;
 import org.gradle.api.internal.tasks.testing.failure.TestFailureMapper;
 import org.gradle.api.internal.tasks.testing.failure.ThrowableToTestFailureMapper;
 import org.gradle.api.tasks.testing.TestFailure;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 /**
- * Maps {@code org.assertj.core.error.MultipleAssertionsError} to {@link TestFailure}.
+ * Maps {@link AssertionError} to {@link TestFailure}.
  * <p>
  * See {@link TestFailureMapper} for more details about failure mapping.
  */
-public class AssertjMultipleAssertionsErrorMapperTest extends TestFailureMapper {
-
+@NonNullApi
+public class AssertErrorMapper extends TestFailureMapper {
     @Override
     protected List<String> getSupportedClassNames() {
         return Collections.singletonList(
-            "org.assertj.core.error.MultipleAssertionsError"
+            AssertionError.class.getName()
         );
     }
 
     @Override
     public TestFailure map(Throwable throwable, ThrowableToTestFailureMapper rootMapper) throws Exception {
-        return TestFailure.fromTestAssertionFailure(throwable, null, null, mapInnerFailures(throwable, rootMapper));
+        return TestFailure.fromTestAssertionFailure(throwable, null, null);
     }
-
-    @SuppressWarnings("unchecked")
-    protected static List<TestFailure> mapInnerFailures(Throwable throwable, ThrowableToTestFailureMapper mapper) throws Exception {
-        List<TestFailure> failures = new ArrayList<TestFailure>();
-
-        List<Throwable> innerExceptions = invokeMethod(throwable, "getErrors", List.class);
-        for (Throwable innerException : innerExceptions) {
-            failures.add(mapper.createFailure(innerException));
-        }
-
-        return failures;
-    }
-
 }
