+++ subprojects/testing-jvm-infrastructure/src/main/java/org/gradle/api/internal/tasks/testing/failure/mappers/AssertjMultipleAssertionsErrorMapper.java
@@ -16,30 +16,43 @@
 
 package org.gradle.api.internal.tasks.testing.failure.mappers;
 
-import org.gradle.api.NonNullApi;
 import org.gradle.api.internal.tasks.testing.failure.TestFailureMapper;
 import org.gradle.api.internal.tasks.testing.failure.ThrowableToTestFailureMapper;
 import org.gradle.api.tasks.testing.TestFailure;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 /**
- * Maps {@link AssertionError} to {@link TestFailure}.
+ * Maps {@code org.assertj.core.error.MultipleAssertionsError} to {@link TestFailure}.
  * <p>
  * See {@link TestFailureMapper} for more details about failure mapping.
  */
-@NonNullApi
-public class AssertErrorMapperTest extends TestFailureMapper {
+public class AssertjMultipleAssertionsErrorMapper extends TestFailureMapper {
+
     @Override
     protected List<String> getSupportedClassNames() {
         return Collections.singletonList(
-            AssertionError.class.getName()
+            "org.assertj.core.error.MultipleAssertionsError"
         );
     }
 
     @Override
     public TestFailure map(Throwable throwable, ThrowableToTestFailureMapper rootMapper) throws Exception {
-        return TestFailure.fromTestAssertionFailure(throwable, null, null);
+        return TestFailure.fromTestAssertionFailure(throwable, null, null, mapInnerFailures(throwable, rootMapper));
     }
+
+    @SuppressWarnings("unchecked")
+    protected static List<TestFailure> mapInnerFailures(Throwable throwable, ThrowableToTestFailureMapper mapper) throws Exception {
+        List<TestFailure> failures = new ArrayList<TestFailure>();
+
+        List<Throwable> innerExceptions = invokeMethod(throwable, "getErrors", List.class);
+        for (Throwable innerException : innerExceptions) {
+            failures.add(mapper.createFailure(innerException));
+        }
+
+        return failures;
+    }
+
 }
