+++ subprojects/testing-jvm-infrastructure/src/main/java/org/gradle/api/internal/tasks/testing/failure/mappers/OpenTestAssertionFailedMapper.java
@@ -32,7 +32,7 @@ import java.util.List;
  * <p>
  * See {@link TestFailureMapper} for more details about failure mapping.
  */
-public class OpenTestAssertionFailedMapperTest extends TestFailureMapper {
+public class OpenTestAssertionFailedMapper extends TestFailureMapper {
 
     @Override
     protected List<String> getSupportedClassNames() {
