+++ subprojects/testing-jvm-infrastructure/src/main/java/org/gradle/api/internal/tasks/testing/failure/mappers/OpenTestMultipleFailuresErrorMapper.java
@@ -29,7 +29,7 @@ import java.util.List;
  * <p>
  * See {@link TestFailureMapper} for more details about failure mapping.
  */
-public class OpenTestMultipleFailuresErrorMapperTest extends TestFailureMapper {
+public class OpenTestMultipleFailuresErrorMapper extends TestFailureMapper {
 
     @Override
     protected List<String> getSupportedClassNames() {
