+++ subprojects/testing-jvm-infrastructure/src/main/java/org/gradle/api/internal/tasks/testing/junit/JUnitTestEventAdapter.java
@@ -24,11 +24,11 @@ import org.gradle.api.internal.tasks.testing.TestResultProcessor;
 import org.gradle.api.internal.tasks.testing.TestStartEvent;
 import org.gradle.api.internal.tasks.testing.failure.DefaultThrowableToTestFailureMapper;
 import org.gradle.api.internal.tasks.testing.failure.TestFailureMapper;
-import org.gradle.api.internal.tasks.testing.failure.mappers.AssertErrorMapperTest;
-import org.gradle.api.internal.tasks.testing.failure.mappers.AssertjMultipleAssertionsErrorMapperTest;
+import org.gradle.api.internal.tasks.testing.failure.mappers.AssertErrorMapper;
+import org.gradle.api.internal.tasks.testing.failure.mappers.AssertjMultipleAssertionsErrorMapper;
 import org.gradle.api.internal.tasks.testing.failure.mappers.JUnitComparisonTestFailureMapper;
-import org.gradle.api.internal.tasks.testing.failure.mappers.OpenTestAssertionFailedMapperTest;
-import org.gradle.api.internal.tasks.testing.failure.mappers.OpenTestMultipleFailuresErrorMapperTest;
+import org.gradle.api.internal.tasks.testing.failure.mappers.OpenTestAssertionFailedMapper;
+import org.gradle.api.internal.tasks.testing.failure.mappers.OpenTestMultipleFailuresErrorMapper;
 import org.gradle.api.tasks.testing.TestFailure;
 import org.gradle.api.tasks.testing.TestResult;
 import org.gradle.internal.id.IdGenerator;
@@ -55,10 +55,10 @@ public class JUnitTestEventAdapter extends RunListener {
 
     private static final List<TestFailureMapper> MAPPERS = Arrays.asList(
         new JUnitComparisonTestFailureMapper(),
-        new OpenTestAssertionFailedMapperTest(),
-        new OpenTestMultipleFailuresErrorMapperTest(),
-        new AssertjMultipleAssertionsErrorMapperTest(),
-        new AssertErrorMapperTest()
+        new OpenTestAssertionFailedMapper(),
+        new OpenTestMultipleFailuresErrorMapper(),
+        new AssertjMultipleAssertionsErrorMapper(),
+        new AssertErrorMapper()
     );
 
     private static final DefaultThrowableToTestFailureMapper FAILURE_MAPPER = new DefaultThrowableToTestFailureMapper(MAPPERS);
