+++ subprojects/testing-jvm-infrastructure/src/test/groovy/org/gradle/api/internal/tasks/testing/failure/mappers/OpenTestMultipleFailuresMapperTestError.groovy
@@ -24,7 +24,7 @@ import spock.lang.Specification
 
 class OpenTestMultipleFailuresMapperTestError extends Specification {
     // SUT
-    def mapper = new OpenTestMultipleFailuresErrorMapperTest()
+    def mapper = new OpenTestMultipleFailuresErrorMapper()
 
     // Our error being mapped
     def error = new MultipleFailuresError(
